+++ API/Router/task.py
@@ -1,46 +1,71 @@
-from fastapi import APIRouter
+from fastapi import APIRouter, Depends
 from schemas import (
-    BaseTask,
+    UpdateUser,
     CreateTask,
     UpdateTask,
     ReadTasksUsingUserId,
     ReadTasksUsingRepoId,
 )
+from typing import Annotated
 from Repository.TaskCRUD import *
+from Authentication.JWTtoken import get_current_user
+from Exceptions import access_denied_not_allowed
 
 router = APIRouter(prefix="/task", tags=["Task"])
 
 
 @router.get("/user_id/{user_id}")
-async def get_tasks_using_user_id(user_id: int) -> list[ReadTasksUsingUserId]:
+async def get_tasks_using_user_id(user_id: int, current_user: Annotated[UpdateUser, Depends(get_current_user)]) -> list[ReadTasksUsingUserId]:
     """The endpoint of getting all tasks using user id"""
+
+    if user_id != current_user.user_id:
+        raise access_denied_not_allowed("read", "tasks")
 
     return await get_tasks_by_user_id(user_id)
 
 
 @router.get("/repo_id/{repo_id}")
-async def get_tasks_using_repo_id(repo_id: int) -> list[ReadTasksUsingRepoId]:
+async def get_tasks_using_repo_id(repo_id: int, current_user: Annotated[UpdateUser, Depends(get_current_user)]) -> list[ReadTasksUsingRepoId]:
     """The endpoint of getting all tasks using repo_id"""
+
+    origin_repo = await check_repo(repo_id)
+
+    if origin_repo.creator_id != current_user.user_id:
+        raise access_denied_not_allowed("read", "tasks")
 
     return await get_tasks_by_repo_id(repo_id)
 
 
 @router.post("/create")
-async def create_task(task: CreateTask) -> None:
+async def create_task(task: CreateTask, current_user: Annotated[UpdateUser, Depends(get_current_user)]) -> None:
     """The endpoint of getting all tasks using repo_id"""
+
+    repo = await check_repo(task.belongs_to_repository_id)
+
+    if (task.creator_id != current_user.user_id) or repo.creator_id != current_user.user_id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="Access denied. You are not allowed to create task for others.")
 
     return await create_new_task(task)
 
 
 @router.put("/update")
-async def update_specific_task(task: UpdateTask) -> None:
+async def update_specific_task(task: UpdateTask, current_user: Annotated[UpdateUser, Depends(get_current_user)]) -> None:
     """The endpoint of updating specific task"""
+
+    repo = await check_repo(task.belongs_to_repository_id)
+
+    if (task.creator_id != current_user.user_id) or repo.creator_id != current_user.user_id:
+        raise access_denied_not_allowed("update", "tasks")
 
     return await update_task_info(task)
 
 
 @router.delete("/delete")
-async def delete_task(task: DeleteTask) -> None:
+async def delete_task(task: DeleteTask, current_user: Annotated[UpdateUser, Depends(get_current_user)]) -> None:
     """The endpoint of deleting specific task"""
+
+    if await find_task_creator(task.task_id) != current_user.user_id:
+        raise access_denied_not_allowed("delete", "tasks")
 
     return await delete_spec_task(task)
